Derive QPACK blocked limit from stream concurrency

+++ include/h2o/qpack.h
@@ -36,7 +36,7 @@ extern const char *h2o_qpack_err_invalid_dynamic_reference;
 extern const char *h2o_qpack_err_invalid_duplicate;
 extern const char *h2o_qpack_err_invalid_pseudo_header;
 
-h2o_qpack_decoder_t *h2o_qpack_create_decoder(uint32_t header_table_size, uint16_t max_blocked);
+h2o_qpack_decoder_t *h2o_qpack_create_decoder(uint32_t header_table_size, uint64_t max_blocked);
 void h2o_qpack_destroy_decoder(h2o_qpack_decoder_t *qpack);
 void h2o_qpack_decoder_cancel_stream(h2o_qpack_decoder_t *qpack, int64_t stream_id);
 /**
@@ -66,7 +66,7 @@ int h2o_qpack_parse_response(h2o_mem_pool_t *pool, h2o_qpack_decoder_t *qpack, i
                              h2o_headers_t *headers, h2o_iovec_t *datagram_flow_id, uint8_t *outbuf, size_t *outbufsize,
                              const uint8_t *src, size_t len, const char **err_desc);
 
-h2o_qpack_encoder_t *h2o_qpack_create_encoder(uint32_t header_table_size, uint16_t max_blocked);
+h2o_qpack_encoder_t *h2o_qpack_create_encoder(uint32_t header_table_size, uint64_t max_blocked);
 void h2o_qpack_destroy_encoder(h2o_qpack_encoder_t *qpack);
 /**
  * Handles packets sent to the QPACK encoder (i.e., the bytes carried by the "decoder" stream)
